% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{HomogeneityOfVariance}
\alias{HomogeneityOfVariance}
\title{Checking homogeneity of variance}
\usage{
HomogeneityOfVariance(inputted.data, dependent.variable, independent.variable)
}
\arguments{
\item{inputted.data}{A dataframe}

\item{dependent.variable}{A string that specifies the column name of the column to use as the dependent variable. Column must be numeric.}

\item{independent.variable}{A string that specifies the column name of the column to use as the independent variable. Column can be numeric or factor. If it's a factor, then it can only have two levels.}
}
\value{
A matrix with two rows. The first row specify what the values are in the second row. The second row:
The first element is the formula used to evaluate levene test. The next element is the p-value from the levene test.
}
\description{
Returns information that can be used
to determine if the homogeneity of variance, an assumption for
ANCOVA, is met.
}
\details{
To test homogeneity of regression slopes, the Levene test is used.
If the p-value is significant, then this means the assumption of homogeneity
of variance is not met.
}
\examples{

dependent.col <- c(10.1, 11.3, 12.1, 13.7, 14.2, 1.6, 2.3, 3.2, 4.1, 5.3)
independent.col <- as.factor(c(1, 1, 1, 1, 1, 0, 0, 0, 0, 0))
covariate.one.col <- c(1, 2, 3, 4, 5, 1, 2, 3, 4, 5)
covariate.two.col <- as.factor(c(1, 0, 1, 0, 1, 0, 1, 0, 1, 0))

inputted.data <- data.frame(dependent.col, independent.col, covariate.one.col,
                            covariate.two.col)

results <- HomogeneityOfVariance(inputted.data, "dependent.col",
                                         "independent.col")

results

}
